Introduce PyDeltaCheckNode
